requirements.txt
allure-pytest==2.13.5
mysql-connector-python==9.0.0
pytest==8.3.2
requests==2.25.1
